feat: extract functions for deploy with proposal
